AA-2003-03-19-1  Bugfix:    Darwin and Freebsd merge from maintrunc                  Changed:   all

+++ source/daemons/qmaster/sge_job.c
@@ -2495,7 +2495,7 @@ int *trigger
       lSetUlong(new_job, JB_priority, uval);
 
       sprintf(SGE_EVENT, MSG_JOB_PRIOSET_SSUI,
-               ruser, rhost, u32c(jobid), uval-BASE_PRIORITY);
+               ruser, rhost, u32c(jobid), (int)(uval-BASE_PRIORITY));
       sge_add_answer(alpp, SGE_EVENT, STATUS_OK, NUM_AN_INFO);
 
    }
